rnd127 - Pseudorandom number generator with period ~2^127
Full 64-bit output from a 32-bit Multiply-With-Carry (MWC)
generator is added with the ouput of a 64-bit Add-with-Carry (AWC)

Tested with the TestU01 library from
http://www.iro.umontreal.ca/~simardr/testu01/tu01.html
Passes all of the tests of BigCrush.
Four 32-bit ranges tested: 64-33, 48-17, 32-1, and 16-49

Criteria for MWCs:
For a MWC with base B and lag L, chose multiplier A such that
A*B^L-1 and A*B^L/2-1 are both prime for a max period of A*B^L/2-1
Using A=~2^32, B=2^32, L=1: Period = A1*2^32/2-1 = ~2^63

The AWC generator only has a lag of 1. On its own it fail any
tests for randomness. The period is 2^64 since the constant being
added is relatively prime with 2^64.

The two generators have periods that are relatively prime with each
other, so the overall period is ~2^63*2^64 = ~2^127

Signed-off-by: James Carter <jwcarter@gmail.com>