Changed way generator state is gotten and set
Generator internal functions: rnd_get_state() and rnd_set_state()
both take an array of uint32_t and either copy the state into the
array or set the state from the array. Both functions expect an
array of the correct size and do no checking.

The library functions rnd_state_to_array() and rnd_array_to_state()
will ensure an array of the correct size is passed in and these
functions should be the normal way of getting and setting state.

Signed-off-by: James Carter <jwcarter@gmail.com>