rnd255 - Pseudorandom number generator with period ~2^255
Full 64-bit output from a 32-bit Multiply-With-Carry (MWC) Lag 7
generator is mixed by multiplying the upper and lower 32-bits
with 32-bit numbers, rotating one of the products by 32-bits
and adding with the other product.

Tested with the TestU01 library from
http://www.iro.umontreal.ca/~simardr/testu01/tu01.html
Passes all of the tests of BigCrush.
Four 32-bit ranges tested: 64-33, 48-17, 32-1, and 16-49

Criteria for MWCs:
For a MWC with base B and lag L, chose multiplier A such that
A*B^L-1 and A*B^L/2-1 are both prime for a max period of A*B^L/2-1
Using A=~2^32, B=2^32, L=7: Period = A1*(2^32)^7/2-1 = ~2^255

Signed-off-by: James Carter <jwcarter@gmail.com>