Better mix the output bits of rnd_126
Flip the upper and lower 32 bits of x2 before adding with x1.
Then multiply the upper and lower 32 bits by two different 32-bit
numbers and add the results after flipping the upper and lower 32
bits of one of the results.

This will make rnd_126 slower, but should improve the randomness.

Signed-off-by: James Carter <jwcarter@gmail.com>