Use lower bits of rnd_127
Have a higher confidence in the lower bits of rnd_127 because the
lower 32 bits of the MWC result is the better random number. The
upper 32 bits of the MWC result are more and more biased towards zero
as it goes from the low to high bits.

Because of this, use the lower 32 and 52 bits when the full 64 bits
are not being returned.

Signed-off-by: James Carter <jwcarter@gmail.com>