Change two of the example 64-bit generators to be like rnd_126
The example generator MWC32 X2 M2 is rnd_126, so modify it to
better mix the output bits like what was done for rnd_126. The
example generator MWC32 x2 is like rnd_126 with out the multiplication
step at the end, so modify it to better mix the output bits as well.

Signed-off-by: James Carter <jwcarter@gmail.com>