rnd_126: Update generator
Do not mix the outputs by rotating the second output by 32-bits and then
adding it to the first output, multiplying the upper and lower 32-bits of
that sum with 32-bit numbers to produce two products, and then rotating the
second product by 32-bits and adding it to the first product.

Just mix by rotating the second output by 32-bits amd adding with the
first output.

Signed-off-by: James Carter <jwcarter@gmail.com>