rnd_127: Changed to a new random number generator
The rand_127 generator now uses a MWC32 lag three generator. The 64-bit
output is mixed by multiplying the upper and lower 32-bits by 32-bit
numbers, rotating the second product by 32-bits and adding it to the
first product to make the new random number.

Passes all of the tests of BigCrush for the following 32-bit ranges:
  03-36, 07-40, 11-44, 15-48, 19-52, 23-56, 27-60, 31-00, 35-04, 39-08,
  43-12, 47-16, 51-20, 55-24, 59-28, 63-32

This generator has a period of ~2^127.

Signed-off-by: James Carter <jwcarter@gmail.com>