Add new 16-sided die test
There are two types of tests. The first takes four random numbers
and creates a 4-bit number (0-15) for each bit position to simulate
a fair sixteen sided die. A count of how many times each side is
rolled is kept and a Chi-square test is performed to determine if the
null hypothesis that all sides are equally probable is evaluated. The
second test is similar but it uses one random number and creates
sixteen 4-bit numbers that are used. The first test is called the
vertical fair die test and the second is called the horizontal fair
die test.

Signed-off-by: James Carter <jwcarter@gmail.com>